Add .gitignore and implement web controller for library management system
- Created .gitignore to exclude build artifacts and temporary files.
- Added main entry point for the application.
- Implemented WebController to manage routes for books, users, and loans.
- Introduced HTML templates for dashboard, books, users, and loans management.
- Established services and repositories for books, users, and loans, enhancing the MVC structure.

+++ internal/users/models/user.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type User struct {
-	ID		int       `json:"ID"`
+	ID		int64     `json:"ID"`
 	Name	string    `json:"name"  binding:"requiredn min=10, max=200"`
 	Email	string    `json:"email" binding:"required.email"`
 	CreatedAt	time.Time `json:"createdAt"`
